+++ Anarchy/DiscordClient.cs
@@ -17,7 +17,8 @@ namespace Discord
             {
                 IEnumerable<string> values;
                 HttpClient.Headers.TryGetValues("Authorization", out values);
-                if (values.ToList().Count > 0) return values.ToList()[0];
+                if (values.ToList().Count > 0)
+                    return values.ToList()[0];
                 else return null;
             }
             set