+++ Anarchy/Guild/Extensions.cs
@@ -112,7 +112,7 @@ namespace Discord
         /// </summary>
         /// <param name="limit">Max amount of guild to receive</param>
         /// <param name="afterId">Guild ID to offset from</param>
-        public static IReadOnlyList<PartialGuild> GetClientGuilds(this DiscordClient client, uint limit = 100, ulong afterId = 0)
+        public static IReadOnlyList<PartialGuild> GetGuilds(this DiscordClient client, uint limit = 100, ulong afterId = 0)
         {
             return client.HttpClient.Get($"/users/@me/guilds?limit={limit}&after={afterId}")
                                 .Deserialize<IReadOnlyList<PartialGuild>>().SetClientsInList(client);